RV32I target reached
unicycle version needs to get revision with LH, LB, SH and SB, everything else was tested.
Multicycle version, Type R and Type I instructions are working. The same goes for LW, SW and BEQ. Needs to get finished
Pipelined version only was tested with LW, SW, OR and BEQ. But, given its nature is very possible that it works as well as the unicycle version